revise and optimize cli subcommand implementations
