+++ cmd/skywire-cli/commands/tp/tp.go
@@ -1,5 +1,5 @@
-// Package clivisor cmd/skywire-cli/commands/visor/transports.go
-package clivisor
+// Package clitransports cmd/skywire-cli/commands/tp/tp.go
+package clitp
 
 import (
 	"bytes"
@@ -11,11 +11,17 @@ import (
 	"text/tabwriter"
 	"time"
 
+	"github.com/bitfield/script"
 	"github.com/google/uuid"
+	"github.com/pterm/pterm"
+	"github.com/pterm/pterm/putils"
 	"github.com/spf13/cobra"
 	"github.com/spf13/pflag"
+	"github.com/tidwall/pretty"
 
 	"github.com/skycoin/skywire-utilities/pkg/cipher"
+	"github.com/skycoin/skywire-utilities/pkg/logging"
+	utilenv "github.com/skycoin/skywire-utilities/pkg/skyenv"
 	clirpc "github.com/skycoin/skywire/cmd/skywire-cli/commands/rpc"
 	"github.com/skycoin/skywire/cmd/skywire-cli/internal"
 	"github.com/skycoin/skywire/pkg/transport"
@@ -27,26 +33,36 @@ var (
 	filterTypes   []string
 	filterPubKeys []string
 	showLogs      bool
+	logger        = logging.MustGetLogger("skywire-cli")
+	removeAll     bool
+	tpTypes       bool
 )
 
 func init() {
-	lsTpCmd.Flags().SortFlags = false
-	RootCmd.AddCommand(tpCmd)
+	tpCmd.Flags().SortFlags = false
+	addTpCmd.Flags().SortFlags = false
+	rmTpCmd.Flags().SortFlags = false
+	discTpCmd.Flags().SortFlags = false
+	treeCmd.Flags().SortFlags = false
 	tpCmd.AddCommand(
-		lsTypesCmd,
-		lsTpCmd,
-		idCmd,
 		addTpCmd,
 		rmTpCmd,
 		discTpCmd,
+		treeCmd,
 	)
+	tpCmd.Flags().StringSliceVarP(&filterTypes, "types", "t", filterTypes, "show transport(s) type(s) comma-separated")
+	tpCmd.Flags().StringSliceVarP(&filterPubKeys, "pks", "p", filterPubKeys, "show transport(s) for public key(s) comma-separated")
+	tpCmd.Flags().BoolVarP(&showLogs, "logs", "l", true, "show transport logs")
+	tpCmd.Flags().StringVarP(&tpID, "id", "i", "", "display transport matching ID")
+	tpCmd.Flags().BoolVarP(&tpTypes, "tptypes", "u", false, "display transport types used by the local visor")
 }
 
 // RootCmd contains commands that interact with the skywire-visor
 var tpCmd = &cobra.Command{
 	Use:   "tp",
-	Short: "View and set transports",
-	Long: `
+	Short: "View and manage transports",
+	Long: `Display and manage transports of the local visor
+
 	Transports are bidirectional communication protocols
 	used between two Skywire Visors (or Transport Edges)
 
@@ -56,68 +72,34 @@ var tpCmd = &cobra.Command{
 	a specific implementation of the Transport.
 
 	Types: stcp stcpr sudph dmsg`,
-}
-
-var lsTypesCmd = &cobra.Command{
-	Use:                   "type",
-	Short:                 "Transport types used by the local visor",
-	Long:                  "\n  Transport types used by the local visor",
-	DisableFlagsInUseLine: true,
 	Run: func(cmd *cobra.Command, _ []string) {
 		rpcClient, err := clirpc.Client(cmd.Flags())
 		if err != nil {
-			os.Exit(1)
+			internal.PrintFatalError(cmd.Flags(), err)
 		}
-		types, err := rpcClient.TransportTypes()
-		internal.Catch(cmd.Flags(), err)
-		internal.PrintOutput(cmd.Flags(), types, fmt.Sprintln(strings.Join(types, "\n")))
-	},
-}
 
-func init() {
-	lsTpCmd.Flags().StringSliceVarP(&filterTypes, "types", "t", filterTypes, "show transport(s) type(s) comma-separated")
-	lsTpCmd.Flags().StringSliceVarP(&filterPubKeys, "pks", "p", filterPubKeys, "show transport(s) for public key(s) comma-separated")
-	lsTpCmd.Flags().BoolVarP(&showLogs, "logs", "l", true, "show transport logs")
-}
+		if tpTypes {
+			types, err := rpcClient.TransportTypes()
+			internal.Catch(cmd.Flags(), err)
+			internal.PrintOutput(cmd.Flags(), types, fmt.Sprintln(strings.Join(types, "\n")))
+			return
+		}
 
-var lsTpCmd = &cobra.Command{
-	Use:   "ls",
-	Short: "Available transports",
-	Long:  "\n    Available transports\n\n    displays transports of the local visor", //nolint
-	Run: func(cmd *cobra.Command, _ []string) {
+		if tpID != "" {
+			tpid := internal.ParseUUID(cmd.Flags(), "transport-id", tpID)
+			tp, err := rpcClient.Transport(tpid)
+			internal.Catch(cmd.Flags(), err)
+			PrintTransports(cmd.Flags(), tp)
+			return
+		}
+
 		var pks cipher.PubKeys
 		if filterPubKeys != nil {
 			internal.Catch(cmd.Flags(), pks.Set(strings.Join(filterPubKeys, ",")))
 		}
-		rpcClient, err := clirpc.Client(cmd.Flags())
-		if err != nil {
-			os.Exit(1)
-		}
 		transports, err := rpcClient.Transports(filterTypes, pks, showLogs)
 		internal.Catch(cmd.Flags(), err)
 		PrintTransports(cmd.Flags(), transports...)
-	},
-}
-
-func init() {
-	idCmd.Flags().StringVarP(&tpID, "id", "i", "", "transport ID")
-}
-
-var idCmd = &cobra.Command{
-	Use:                   "id (-i) <transport-id>",
-	Short:                 "Transport summary by id",
-	Long:                  "\n    Transport summary by id", //nolint
-	DisableFlagsInUseLine: true,
-	Run: func(cmd *cobra.Command, args []string) {
-
-		tpid := internal.ParseUUID(cmd.Flags(), "transport-id", args[0])
-		rpcClient, err := clirpc.Client(cmd.Flags())
-		if err != nil {
-			os.Exit(1)
-		}
-		tp, err := rpcClient.Transport(tpid)
-		internal.Catch(cmd.Flags(), err)
-		PrintTransports(cmd.Flags(), tp)
 	},
 }
 
@@ -134,7 +116,7 @@ func init() {
 }
 
 var addTpCmd = &cobra.Command{
-	Use:                   "add (-p) <remote-public-key>",
+	Use:                   "add",
 	Short:                 "Add a transport",
 	Long:                  "\n    Add a transport\n    \n    If the transport type is unspecified,\n    the visor will attempt to establish a transport\n    in the following order: skywire-tcp, stcpr, sudph, dmsg",
 	Args:                  cobra.MinimumNArgs(1),
@@ -198,7 +180,7 @@ func init() {
 }
 
 var rmTpCmd = &cobra.Command{
-	Use:                   "rm ( -a || -i ) <transport-id>",
+	Use:                   "rm",
 	Short:                 "Remove transport(s) by id",
 	Long:                  "\n    Remove transport(s) by id",
 	DisableFlagsInUseLine: true,
@@ -278,7 +260,7 @@ func init() {
 }
 
 var discTpCmd = &cobra.Command{
-	Use:                   "disc (--id=<transport-id> || --pk=<edge-public-key>)",
+	Use:                   "disc",
 	Short:                 "Discover remote transport(s)",
 	Long:                  "\n    Discover remote transport(s) by ID or public key",
 	DisableFlagsInUseLine: true,
@@ -360,4 +342,172 @@ func (t *transportID) Set(s string) error {
 	}
 	*t = transportID(tID)
 	return nil
+}
+
+var (
+	sortedEdgeKeys []string
+	utURL          string
+	tpdURL         string
+	cacheFileTPD   string
+	cacheFileUT    string
+	cacheFilesAge  int
+	padSpaces      int
+	isStats        bool
+	rawData        bool
+	refinedData    bool
+	noFilterOnline bool
+)
+
+// RootCmd is tpCmd
+var RootCmd = tpCmd
+
+func init() {
+	treeCmd.Flags().StringVarP(&tpdURL, "tpdurl", "a", utilenv.TpDiscAddr, "transport discovery url")
+	treeCmd.Flags().StringVarP(&utURL, "uturl", "w", utilenv.UptimeTrackerAddr, "uptime tracker url")
+	treeCmd.Flags().BoolVarP(&rawData, "raw", "r", false, "print raw json data")
+	treeCmd.Flags().BoolVarP(&refinedData, "pretty", "p", false, "print pretty json data")
+	treeCmd.Flags().BoolVarP(&noFilterOnline, "noton", "o", false, "do not filter by online status in UT")
+	treeCmd.Flags().StringVar(&cacheFileTPD, "cft", os.TempDir()+"/tpd.json", "TPD cache file location")
+	treeCmd.Flags().StringVar(&cacheFileUT, "cfu", os.TempDir()+"/ut.json", "UT cache file location.")
+	treeCmd.Flags().IntVarP(&cacheFilesAge, "cfa", "m", 5, "update cache files if older than n minutes")
+	//TODO: calculate tree levels initially and apply appropriate padding ; as an alternative to manually padding
+	treeCmd.Flags().IntVarP(&padSpaces, "pad", "P", 15, "padding between tree and tpid")
+	treeCmd.Flags().BoolVarP(&isStats, "stats", "s", false, "return only statistics")
+}
+
+var treeCmd = &cobra.Command{
+	Use:   "tree",
+	Short: "tree map of transports on the skywire network",
+	Long:  fmt.Sprintf("display a tree representation of transports from TPD\n\n%v/all-transports\n\nSet cache file location to \"\" to avoid using cache files", utilenv.TpDiscAddr),
+	Run: func(cmd *cobra.Command, args []string) {
+		tps := internal.GetData(cacheFileTPD, tpdURL+"/all-transports", cacheFilesAge)
+		if rawData {
+			script.Echo(tps).Stdout() //nolint
+			return
+		}
+		if refinedData {
+			script.Echo(string(pretty.Color(pretty.Pretty([]byte(tps)), nil))).Stdout() //nolint
+			return
+		}
+		var uts string
+		var utkeys []string
+		var offlinekeys []string
+		if !noFilterOnline {
+			uts = internal.GetData(cacheFileUT, utURL+"/uptimes?v=v2", cacheFilesAge)
+			utkeys, _ = script.Echo(uts).JQ(".[] | select(.on) | .pk").Replace("\"", "").Slice()             //nolint
+			offlinekeys, _ = script.Echo(uts).JQ(".[] | select(.on  | not) | .pk").Replace("\"", "").Slice() //nolint
+		}
+
+		sortedEdgeKeys, _ = script.Echo(tps).JQ(".[].edges[]").Freq().Column(2).Slice() //nolint
+
+		if isStats {
+			fmt.Printf("Unique keys in Transport Discovery: %d\n", len(sortedEdgeKeys))
+			tpcount, _ := script.Echo(tps).JQ(".[].type").CountLines() //nolint
+			fmt.Printf("Count of transports: %v\n", tpcount)
+			tptypes, _ := script.Echo(tps).JQ(".[].type").Freq().String() //nolint
+			fmt.Printf("types of transports: \n%v\n", tptypes)
+			vcount, _ := script.Echo(tps).JQ(".[].edges[]").Freq().String() //nolint
+			fmt.Printf("Visors by transport count:\n%v\n", vcount)
+			return
+		}
+
+		fmt.Printf("Tree        *Online        %s        %s                            TPID                                 Type\n", pterm.Black(pterm.BgRed.Sprint("*Offline")), pterm.Red("*Not in UT"))
+
+		leveledList := pterm.LeveledList{}
+		edgeKey := sortedEdgeKeys[0]
+		leveledList = append(leveledList, pterm.LeveledListItem{Level: 0, Text: filterOnlineStatus(utkeys, offlinekeys, edgeKey)})
+
+		var usedkeys []string
+		usedkeys = append(usedkeys, edgeKey)
+		var lvl func(n int, k string)
+		lvl = func(n int, k string) {
+			l, _ := script.Echo(tps).JQ(".[] | select(.edges[] == " + k + ") | .edges[] | select(. != " + k + ")").Slice() //nolint
+			for _, m := range l {
+				if m == k {
+					continue
+				}
+				var ok bool
+				ok = false
+				for _, o := range usedkeys {
+					if m == o {
+						ok = true
+					}
+				}
+				if ok {
+					continue
+				}
+				usedkeys = append(usedkeys, m)
+				var tpid string
+				if n == 0 {
+					tpid = ""
+				} else {
+					tpid, _ = script.Echo(tps).JQ(".[] | select(.edges | index(" + k + ") and index(" + m + ")) | .t_id + \" \" + .type").First(1).String() //nolint
+				}
+				leveledList = append(leveledList, pterm.LeveledListItem{Level: n, Text: strings.ReplaceAll(strings.ReplaceAll(fmt.Sprintf("%s %s", filterOnlineStatus(utkeys, offlinekeys, m), strings.Repeat(" ", func() int {
+					indent := padSpaces - 4 - n*2
+					if indent < 0 {
+						return 0
+					}
+					return indent
+				}())+tpid), "\n", ""), "\"", "")})
+				lvl(n+1, m)
+			}
+		}
+		lvl(1, edgeKey)
+
+		pterm.DefaultTree.WithRoot(putils.TreeFromLeveledList(leveledList)).Render() //nolint
+		for _, edgeKey := range sortedEdgeKeys {
+			found := false
+			for _, usedKey := range usedkeys {
+				if usedKey == edgeKey {
+					found = true
+					break
+				}
+			}
+			if !found {
+				leveledList = pterm.LeveledList{}
+				leveledList = append(leveledList, pterm.LeveledListItem{Level: 0, Text: filterOnlineStatus(utkeys, offlinekeys, edgeKey)})
+				usedkeys = append(usedkeys, edgeKey)
+				lvl(1, edgeKey)
+				pterm.DefaultTree.WithRoot(putils.TreeFromLeveledList(leveledList)).Render() //nolint
+			}
+		}
+		l, _ := script.Echo(tps).JQ(".[] | select(.edges[0] == .edges[1]) | .edges[0] + \""+strings.Repeat(" ", padSpaces)+"\" + .t_id + \" \" + .type").Replace("\"", "").Slice() //nolint
+		if len(l) > 0 {
+			pterm.Println(pterm.Red("Self-transports"))
+			for _, m := range l {
+				pterm.Println(filterOnlineStatus(utkeys, offlinekeys, m))
+			}
+		}
+	},
+}
+
+func filterOnlineStatus(utkeys, offlinekeys []string, key string) (lvlN string) {
+	isOnline, isOffline := false, false
+	lvlN = strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(key, " ", ""), "\t", ""), "\n", ""), "\"", "")
+	if !noFilterOnline {
+		for _, k := range utkeys {
+			if strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(k, " ", ""), "\t", ""), "\n", ""), "\"", "") == strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(key, " ", ""), "\t", ""), "\n", ""), "\"", "") {
+				isOnline = true
+				break
+			}
+		}
+		if !isOnline {
+			for _, k := range offlinekeys {
+				if strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(k, " ", ""), "\t", ""), "\n", ""), "\"", "") == strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(strings.ReplaceAll(key, " ", ""), "\t", ""), "\n", ""), "\"", "") {
+					isOffline = true
+					break
+				}
+			}
+		}
+	} else {
+		isOnline, isOffline = true, false
+	}
+	if !isOnline && !isOffline {
+		lvlN = pterm.Red(strings.ReplaceAll(key, "\"", ""))
+	}
+	if isOffline {
+		lvlN = pterm.Black(pterm.BgRed.Sprint(strings.ReplaceAll(key, "\"", "")))
+	}
+	return lvlN
 }
